Add option disableAnsi to color converters
Option `disableAnsi` turns off ANSI color output in the
HighlightConverter and StyleConverter unconditionally. The existing
option `noConsoleNoAnsi` option only takes effect if System.console() is
null, which is not always the case. An example of this is when running
an application that writes to standard output inside a container or a
service wrapper.